Combine an Array into a String Using the join Method **
The join method is used to join the elements of an array together to create a string. It takes an argument for the delimiter that is used to separate the array elements in the string.

Here's an example:

var arr = ["Hello", "World"];
var str = arr.join(" ");
str would have a value of the string Hello World.

Use the join method (among others) inside the sentensify function to make a sentence from the words in the string str. The function should return a string. For example, I-like-Star-Wars would be converted to I like Star Wars. For this challenge, do not use the replace method.

Answer:
function sentensify(str) {
  // Only change code below this line
  var newArr = str.split(/\W/).join(" ");
  return newArr;
  // Only change code above this line
}

console.log(sentensify("May-the-force-be-with-you"));