Sum All Numbers in a Range ***
We'll pass you an array of two numbers. Return the sum of those two numbers plus the sum of all the numbers between them. The lowest number will not always come first.

For example, sumAll([4,1]) should return 10 because sum of all the numbers between 1 and 4 (both inclusive) is 10.

Answer:
function sumAll(arr) {
  //Since numbers won't always be in order, max and min will help organize
  let max = Math.max(arr[0], arr[1]); //create variable to store max number
  let min = Math.min(arr[0], arr[1]); //create variable to store min number
  let sumBetween = 0; //create accumulator variable to hold incrememented sum
  for (let i = min; i <= max; i++) {
    sumBetween += i;
  }
  return sumBetween;
}

sumAll([1, 4]);